pyproject.toml
[tool.poetry]
name = "telegram-f1-bot"
version = "0.1.0"
description = ""
authors = ["Riccardo Trivellato <6997999+rtrive@users.noreply.github.com>", "Marco Comi <marco.comi@aol.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.3.0"
python-dotenv = "^0.20.0"
undetected-chromedriver = "^3.1.5"
selenium-wire = "^4.6.4"
requests = "^2.27.1"
prettytable = "^3.3.0"
psutil = "^5.9.1"
python-telegram-bot = "^13.12"
APScheduler="^3.6.3"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.37"
pyopenssl = "22.0.0"
cryptography = "38.0.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.960"
types-requests = "^2.27.31"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
